// src/types/theme/theme.ts
export interface ColorThemeTypes {
  font100: string;
  font200: string;
  font300: string;
  font400: string;
  font500: string;
  font600: string;
  bg100: string;
  bg200: string;
  bg300: string;
  bg400: string;
  bg500: string;
  accent100: string;
  accent200: string;
  accent300: string;
  accent400: string;
  accent500: string;
  accent600: string;
  accent700: string;
  accent800: string;
  accent900: string;
  accent1000: string;
  border100: string;
  border200: string;
}

export interface Theme {
  color: ColorThemeTypes;
}